Relative slider movement
Use relative touch dragging instead of absolute position taps, this makes it easier to perform slight adjustments to a slider value.
Some bugfixes, now the slider value is between 0.0 and 1.0 inclusive whereas before in could never get completely to zero or one.
Some cleanup. The touch is now tracked, additional touches to the same slider are rejected.
Use weakSelf in the completion callback for animateWithDuration.